pyproject.toml
[tool.poetry]
name = "cfn_kafka_admin"
version = "0.6.0-rc2"
description = "AWS CloudFormation Resources to manage Kafka"
authors = ["John Mille <john@compose-x.io>"]
classifiers = [
  "License :: OSI Approved :: Mozilla Public License 2.0 (MPL 2.0)",
  "Natural Language :: English",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
]

readme = "README.rst"
license = "MPL-2.0"

[tool.poetry.urls]
"Source" = "https://github.com/compose-x/cfn-kafka-admin"
"Bug Tracker" = "https://github.com/compose-x/cfn-kafka-admin/issues"

[tool.poetry.dependencies]
python = "^3.10"
compose-x-common = "^1.4"
jsonschema = "^4.21"
importlib-resources = "^6.0"
PyYAML = "^6.0"
troposphere = "^4.5.1"
kafka-schema-registry-admin = "^0.5"
datamodel-code-generator = { extras = ["http"], version = "^0.25" }
aws-cfn-custom-resource-resolve-parser = "^0.3"
cfn-resource-provider = "^1.0.7"
retry2 = "^0.9.5"
confluent-kafka = "^2.3"

[tool.poetry.scripts]
aws-cfn-kafka-admin-provider = "cfn_kafka_admin.cli:main"

[tool.poetry.group.dev.dependencies]
black = "^23.7"
isort = "^5.10"
coverage = "^5.5"
pytest = "^7.4"
pre-commit = "^2.19.0"
tbump = "^6.8.0"
cleanpy = "^0.3.1"
Sphinx = "^4.5.0"
sphinx-material = "^0.0.35"
sphinx-jsonschema = "^1.19.1"
pyupgrade = "^3.3.1"
testcontainers = "^4.0"

[tool.datamodel-codegen]
input = "cfn_kafka_admin/specs/aws-cfn-kafka-admin-provider-schema.json"
input-file-type = "jsonschema"
output = "cfn_kafka_admin/models/admin.py"
#output-model-type = "dataclasses.dataclass"
reuse-model = "true"
target-python-version = "3.10"
use-schema-description = "true"
use-field-description = "true"
use-double-quotes = "true"
disable-timestamp = "true"


[tool.tbump]
github_url = "https://github.com/compose-x/cfn-kafka-admin"

[tool.tbump.version]
current = "0.6.0-rc2"

regex = '''
  (?P<major>\d+)
  \.
  (?P<minor>\d+)
  \.
  (?P<patch>\d+)
  (?:(?P<rc>[\S]+))?
  '''

[tool.tbump.git]
message_template = "Bump to {new_version}"
tag_template = "v{new_version}"

[[tool.tbump.file]]
src = "pyproject.toml"

[[tool.tbump.file]]
src = "cfn_kafka_admin/__init__.py"

[[tool.tbump.before_commit]]
name = "Files format"
cmd = "make conform"

[tool.black]
line-length = 88

[tool.isort]
profile = "black"
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
known_first_party = "kelvin"

[tool.coverage.report]
exclude_lines = [
  "if __name__ == '__main__'"
]
[tool.coverage.run]
omit = [
  "*/cli.py"
]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
